Pin ReactiveSearch to a functional release; memoize/optimize rendering of Search Results for general Search and Collection search
